+++ bitcointx/core/serialize.py
@@ -77,6 +77,22 @@ class DeserializationExtraDataError(SerializationError):
         self.padding = padding
 
 
+class DeserializationValueBoundsError(SerializationError):
+    """Deserialized value out of bounds
+
+    Thrown by deserialize() when a deserialized value turns out to be out
+    of allowed bounds
+    """
+
+    def __init__(self, msg: str, *, klass: Type['Serializer[T_unbounded]'],
+                 value: int, upper_bound: int, lower_bound: int):
+        super().__init__(msg)
+        self.klass = klass
+        self.value = value
+        self.upper_bound = upper_bound
+        self.lower_bound = lower_bound
+
+
 def ser_read(f: ByteStream_Type, n: int) -> bytes:
     """Read from a stream safely
 
@@ -185,6 +201,7 @@ class ImmutableSerializable(Serializable):
             object.__setattr__(self, '_cached__hash__', _cached__hash__)
             return _cached__hash__
 
+
 class Serializer(Generic[T_unbounded]):
     """Base class for object serializers"""
     def __new__(cls: Type['Serializer[T_unbounded]']
@@ -236,14 +253,52 @@ class VarIntSerializer(Serializer[int]):
     @classmethod
     def stream_deserialize(cls, f: ByteStream_Type, **kwargs: Any) -> int:
         r = ser_read(f, 1)[0]
+
         if r < 0xfd:
             return r
-        elif r == 0xfd:
-            return int(struct.unpack(b'<H', ser_read(f, 2))[0])
+
+        if r == 0xfd:
+            v = int(struct.unpack(b'<H', ser_read(f, 2))[0])
+            lower_bound = 0xfd
+            if v < lower_bound:
+                raise DeserializationValueBoundsError(
+                    f"non-canonical 3-byte compact size for variable integer: "
+                    f"0x{v:x} less than 0x{lower_bound:x}",
+                    klass=cls, value=v, lower_bound=lower_bound,
+                    upper_bound=0xFFFF)
         elif r == 0xfe:
-            return int(struct.unpack(b'<I', ser_read(f, 4))[0])
+            v = int(struct.unpack(b'<I', ser_read(f, 4))[0])
+
+            lower_bound = 0x10000
+            if v < lower_bound:
+                raise DeserializationValueBoundsError(
+                    f"non-canonical 5-byte compact size for variable integer: "
+                    f"0x{v:x} less than 0x{lower_bound:x}",
+                    klass=cls, value=v, lower_bound=lower_bound,
+                    upper_bound=0xFFFFFFFF)
         else:
-            return int(struct.unpack(b'<Q', ser_read(f, 8))[0])
+            v = int(struct.unpack(b'<Q', ser_read(f, 8))[0])
+
+            lower_bound = 0x100000000
+            if v < lower_bound:
+                raise DeserializationValueBoundsError(
+                    f"non-canonical 9-byte compact size for variable integer: "
+                    f"0x{v:x} less than 0x{lower_bound:x}",
+                    klass=cls, value=v, lower_bound=lower_bound,
+                    upper_bound=MAX_SIZE)
+
+        # With MAX_SIZE being defined as less than 32-bit max value,
+        # this means that any canonically encoded 64-bit value will be
+        # more than MAX_SIZE. This also means that upper_bound supplied
+        # to the exception may happen to be less than lower bound.
+        if v > MAX_SIZE:
+            raise DeserializationValueBoundsError(
+                f"non-canonical compact size for variable integer: "
+                f"0x{v:x} more than 0x{MAX_SIZE:x}",
+                klass=cls, value=v, lower_bound=lower_bound,
+                upper_bound=MAX_SIZE)
+
+        return v
 
 
 class BytesSerializer(Serializer[bytes]):